Change logout to POST and redirect to session
I am learning postman and it was confusing getting redirected to "/"
and being informed that I need to specify a resource on GET request
when I was sending a POST request. It really felt like a failed
POST. I like this behavior better for an API, but I don't know what
the correct behavior is here.